Split OTEL Config package files
